refactor: [step2] LottoShop 에서 Lottos 분리
